Extract reusable feed store specs helper methods into a shared scope so it can be reused by any other feed store implementation tests
